requirements.txt

Pillow==7.1.2

streamlit==1.9.2

tensorflow==2.8.0+zzzcolab20220506162203

